refactor: extract student dashboard presentation and logic
