+++ src/components/siswa/index.ts
@@ -1,0 +1,11 @@
+export { StudentDashboardHeader } from "./StudentDashboardHeader";
+export type { StudentDashboardHeaderProps } from "./StudentDashboardHeader";
+
+export { StudentStatsCards } from "./StudentStatsCards";
+export type { StudentStatsCardsProps } from "./StudentStatsCards";
+
+export { StudentGradesTable } from "./StudentGradesTable";
+export type { StudentGradesTableProps } from "./StudentGradesTable";
+
+export { StudentAttendanceTable } from "./StudentAttendanceTable";
+export type { StudentAttendanceTableProps } from "./StudentAttendanceTable";
